made changes per review
