+++ CMakeLists.txt
@@ -75,11 +75,8 @@ set(kvstore_dir src/key_value_store/)
 add_executable(caw_cli src/cli_tool/caw_cli.cpp)
 target_link_libraries(caw_cli gflags::gflags)
 
-add_executable(kv_store src/key_value_store/kv_store.cpp ${store_proto_srcs} ${store_grpc_srcs})
+add_executable(kv_store src/key_value_store/kv_store.cpp ${kvstore_dir}/server/kvstore_server.cpp ${kvstore_dir}/core/KeyValueStore.cpp ${store_proto_srcs} ${store_grpc_srcs})
 target_link_libraries(kv_store ${_GRPC_GRPCPP} ${_PROTOBUF_LIBPROTOBUF} gflags::gflags glog::glog)
-
-add_executable(kv_store_runner ${kvstore_dir}/server/kvstore_server_runner.cpp ${kvstore_dir}/server/kvstore_server.cpp ${kvstore_dir}/core/KeyValueStore.cpp ${store_proto_srcs} ${store_grpc_srcs})
-target_link_libraries(kv_store_runner ${_GRPC_GRPCPP} ${_PROTOBUF_LIBPROTOBUF} gflags::gflags glog::glog)
 
 add_executable(kv_store_test tests/KeyValueStoreTests.cpp src/key_value_store/core/KeyValueStore.cpp)
 target_link_libraries(kv_store_test gtest gtest_main)
